Ensure the archive is extracted
During my run, I got to this point without the tar being fully extracted. It was supposed to be extracted by one of the Crucible calls, but I was re-running Crucible so much and touching it that it skipped it at one point.

Just to be certain, folks should always extract here.. until we're content with doing it elsewhere or have a strong enough argument not to anymore.